+++ packages/blade-core/src/utils/lodashButBetter/capitalize.ts
@@ -1,0 +1,5 @@
+const capitalize = <T extends string>(str: T): string => {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+};
+
+export { capitalize };