+++ packages/blade-core/src/utils/lodashButBetter/keys.ts
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+function keys<T extends Record<string, any>>(obj: T): Array<keyof T> {
+  return Object.keys(obj) as Array<keyof T>;
+}
+
+export default keys;