+++ packages/blade-core/src/utils/makeAccessible/index.ts
@@ -1,0 +1,2 @@
+export * from './makeAccessible';
+export * from './types';