+++ packages/blade-svelte/src/components/Button/index.ts
@@ -1,0 +1,4 @@
+export { default as Button } from './Button.svelte';
+export type { ButtonProps } from './types';
+export { BaseButton } from './BaseButton';
+export type { BaseButtonProps } from './BaseButton';