+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @razorpay/blade
 
+## 12.75.0
+
+### Minor Changes
+
+- 6f80d8b54: feat(blade): add add support to modify dataKey and nameKey in donut chart
+
+### Patch Changes
+
+- d9f9ef9ba: ci: update npm publishing via OIDC
+- 297cd9469: fix(blade): update ci bot permissions
+- 4f33fce21: fix: update storybook example & expose ChartLabelContentProps type
+
+## 12.74.0
+
+### Minor Changes
+
+- 3498bd73f: feat(blade): added GenUI components
+
+## 12.73.1
+
+### Patch Changes
+
+- 7320c7e4b: feat(blade): patch for table pagination
+
+## 12.73.0
+
+### Minor Changes
+
+- 16240ddd4: fix(blade): default pagesize in pagination
+
+## 12.72.0
+
+### Minor Changes
+
+- 069a7163d: Releasing Blade components to Svelte with new packages Blade-core and Blade-svelte
+- ac14ceb6a: Version mismatch issue fixed for blade svelte release
+
+## 12.71.0
+
+### Minor Changes
+
+- b6c2e2563: feat(blade): expose tick formatter
+
 ## 12.70.1
 
 ### Patch Changes